Move SideData to separate class
